Allow prefixes to be specified statically in the config
If I want to use gunicorn with multiple workers to run this app in
production, having each worker generate their own specific prefixes is
going to cause a headache!

Whilst there is probably a nicer solution to share the prefixes between
the workers in the pool, allowing the prefixes to be statically declared
in the config file also works and is sufficient for my needs.